Updated MenDBib to v1.4
Turned out that the LaTeX compiler was unhappy with comments stuffed inside the bibtex entries (comments were put there when an entry had a journal name that was not found in the hardwired abbreviation look-up table).  New format:  leave out those comments and put them into a separate (non-bib) file called "mendbib_info.README".  This readme file will also contain the big block of description that was originally at the beginning of the mendbib.bib file, as well as a statement of the number of entries in mendbib.bib and in mendbib_prob.bib.  Mendbib.bib is now cleaner, with just 2 commented lines at the top (a "%top" that is needed by the code, and a second comment giving the time stamp when the file was generated), and then nothing but pure bibtex entries.  I think this sleeker appearance will avoid the LaTeX compilation problems I was having.  Oh, one more thing:  in the case of an unrecognized journal for which a look-up abbreviation could not be found, the strategy for putting an abbreviation into one's LaTeX document has changed somewhat.  Turns out that putting a \newcommand line near the top of one's latex paper (typically where there is a block of other \newcommand lines to define one's personal preferences) works.  The @preamble line into a second bibtex file did not work.  This new way is much more straightforward.  Instructions printed to the mendbib_info.README regarding how to deal with unrecognized journals have changed accordingly.