Updated MenDBib to v1.5
Just a minor (but important) tweak to the code:  the check to determine if there were 2 or more identical entries was originally flawed, as there could be insignificant differences in the bibtex entry content that would make 2 entries that were the same appear to be different (for example, in one entry the author list might include a "L. W" and in the other entry the list might show as "L.~W.").  The comparison is now performed against the citeKey rather than comparing one whole bibtex entry to another. The citeKeys will be identical for 2 entries that are referencing the same article, even if there are minor differences in the way that the reference materials are written.