Updated MenDBib to V1.6
Inserted a window refresh following successful download of bibtex to Dropbox.  A refresh of the ShareLaTeX window appears necessary in order to clear the cache of possibly obsolete citation information and to allow ShareLaTeX to "see" the new bibtex file.  Also a few new aesthetic changes (the MenDBib icon now appears, and the user button has been bumped up from the bottom of the window a bit more to make it easier to see.