updated MenDBib to version 2.0
Changed the way that MenDBib normalizes journal names.  The end-product should be the same as it was in the previous version, but I have re-written  the function "fixJournal" so that adding journal names to the list should be greatly simplified in contrast to the previous version, where some really ugly RegEx was used to do the parsing.  In re-writing fixJournal, some other relatively minor changes needed to be made elsewhere in the code.  This new version is not adding capability, but rather simplifying the way that journal names are normalized and turned into abbreviations.